DMML_Assignment_Customer_Churn_Pipeline-main/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
amqp==5.1.1
annotated-types==0.7.0
anyio==3.6.2
apache-airflow==2.6.1
apache-airflow-providers-celery==3.1.0
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.2
apispec==5.2.2
argcomplete==3.0.8
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
bigtree==0.25.2
billiard==3.6.4.0
blinker==1.6.2
cachelib==0.9.0
cattrs==22.2.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
cloudpickle==3.1.1
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
contourpy==1.3.0
cron-descriptor==1.3.0
croniter==1.3.14
cryptography==40.0.2
cycler==0.12.1
dask==2024.8.0
dask-expr==1.1.10
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.3.0
docutils==0.20
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.1.1
fastapi==0.115.11
feast==0.47.0
Flask==2.2.5
Flask-AppBuilder==4.3.0
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
flower==1.2.0
fonttools==4.56.0
frozenlist==1.3.3
fsspec==2025.3.0
graphviz==0.20.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httptools==0.6.4
httpx==0.23.3
humanize==4.6.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.17.3
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kombu==5.2.4
lazy-object-proxy==1.9.0
limits==3.4.0
linkify-it-py==2.0.2
locket==1.0.0
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.9.4
mdit-py-plugins==0.3.5
mdurl==0.1.2
mmh3==5.1.0
multidict==6.0.4
mypy==1.15.0
mypy-extensions==1.0.0
numpy==1.26.4
openpyxl==3.1.5
ordered-set==4.1.0
packaging==21.3
pandas==2.2.3
partd==1.4.2
pathspec==0.9.0
pendulum==2.1.2
pillow==11.1.0
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==6.30.1
psutil==5.9.5
pyarrow==18.0.0
pycparser==2.21
pydantic==1.10.1
pydantic_core==2.27.2
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.3.5
rich_argparse==1.1.0
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlalchemy2-stubs==0.0.2a38
sqlparse==0.4.4
starlette==0.46.1
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
tornado==6.3.2
tqdm==4.67.1
typeguard==4.4.2
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==1.26.15
uvicorn==0.34.0
uvicorn-worker==0.3.0
uvloop==0.21.0
vine==5.0.0
watchfiles==1.0.4
wcwidth==0.2.6
websockets==15.0.1
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.15.0